% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidap.R
\name{fidap_cleint}
\alias{fidap_cleint}
\title{fidap_client}
\usage{
fidap_cleint(api_key, source)
}
\arguments{
\item{api_key}{place your dashboard api_key here.}

\item{source}{Source is bq(bigquery), sf_aws(Snowflake AWS) or sf_gcp(Snowflake GCP)}
}
\value{
A object
}
\description{
fidap_client
}
