% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidap.R
\name{fidap_client}
\alias{fidap_client}
\title{fidap_client}
\usage{
fidap_client(api_key, source)
}
\arguments{
\item{api_key}{place your dashboard api_key here.}

\item{source}{Source is bq(bigquery), sf_aws(Snowflake AWS) or sf_gcp(Snowflake GCP)}
}
\value{
A object
}
\description{
fidap_client tasks two arguments api_key and source, Api_key can be get from app.fidap.com, It returns class Object which contains $sql method, this method allow user to query the data via fidap.
}
